requirements.txt

# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3

# Database
psycopg2-binary==2.9.11
redis==5.0.1
django-redis==5.4.0

# Authentication
djangorestframework-simplejwt==5.3.0
social-auth-app-django==5.4.0
PyJWT==2.8.0

# WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# AWS
boto3==1.29.7
django-storages==1.14.2

# Celery for background tasks
celery==5.3.4
django-celery-beat==2.5.0

# Utilities
python-decouple==3.8
Pillow==11.3.0
markdown==3.5.1
bleach==6.1.0
python-slugify==8.0.1

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
